+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### FUTURE
 
+* [ADDED] new `$is: Boolean` option
 * [ADDED] `keywordPrefix` validator option - makes it possible to change default schema definition keyword prefix (`$`)
 * [ADDED] Browser support
 * [FIXED] Resolving of `$or` conditions - see [#2](https://github.com/fogine/json-inspector/issues/2)